+++ status/api/views.py
@@ -18,6 +18,7 @@ class JSONResponseMixin(object):
     """
     A mixin that can be used to render a JSON response.
     """
+
     def render_to_json_response(self, context, **response_kwargs):
         """
         Returns a JSON response, transforming 'context' to make the payload.
@@ -38,6 +39,7 @@ class JSONView(JSONResponseMixin, View):
     """
     View that returns a JSON response.
     """
+
     def render_to_response(self, context, **response_kwargs):
         return self.render_to_json_response(context, **response_kwargs)
 
@@ -46,6 +48,7 @@ class APIView(JSONView, ContextMixin):
     """
     Base class for API views.
     """
+
     def get(self, request, *args, **kwargs):
         context = self.get_context_data(**kwargs)
         return self.render_to_response(context)
@@ -64,12 +67,84 @@ class RootAPIView(APIView):
     """
     Root API view that routes to each single ProviderAPIView
     """
+
     def __init__(self):
-        self.providers = settings.CHECK_PROVIDERS
         super(RootAPIView, self).__init__()
+        self.providers = settings.PROVIDERS
+
+    def _get_provider_url(self, request, resource, name):
+        return request.build_absolute_uri(reverse('api_{}_{}'.format(resource, name)))
 
     def get(self, request, *args, **kwargs):
-        context = {name: request.build_absolute_uri(reverse("api_{}".format(name))) for name, _, _, _ in self.providers}
+        """
+        Create an object that contains all resources and within them, all providers with their current status.
+
+        Mapping of resource to another mapping of provider to status.
+        Resource -> {Provider -> Status}
+
+        :param request: Request.
+        :return: Rendered response.
+        """
+        context = {resource: {name: ProviderAPIView(provider, args, kwargs).get_context_data()
+                              for name, provider, args, kwargs in providers}
+                   for resource, providers in self.providers.items()}
+        return self.render_to_response(context)
+
+    def options(self, request, *args, **kwargs):
+        """
+        List all providers of current resource with their absolute urls for each resource
+
+        Mapping of resource to another mapping of provider to url.
+        Resource -> {Provider -> URL}
+
+        :param request: Request.
+        :return: Rendered response.
+        """
+        context = {resource: {name: self._get_provider_url(request, resource, name) for name, _, _, _ in providers}
+                   for resource, providers in self.providers.items()}
+        return self.render_to_response(context)
+
+
+class ResourceAPIView(RootAPIView):
+    """
+    Root API view for a resource. List all checks
+    """
+    resource = None
+
+    def __init__(self, resource, **kwargs):
+        super(ResourceAPIView, self).__init__()
+        self.resource = resource
+        try:
+            self.providers = settings.PROVIDERS[resource]
+        except KeyError:
+            raise ValueError("Resource doesn't exists: %s" % (resource,))
+
+    def get(self, request, *args, **kwargs):
+        """
+        Create an object that contains all providers of current resource with their current status.
+
+        Mapping of provider to status.
+        Provider -> Status
+
+        :param request: Request.
+        :return: Rendered response.
+        """
+        context = {name: ProviderAPIView(provider=p, provider_args=args, provider_kwargs=kwargs).get_context_data()
+                   for name, p, args, kwargs in self.providers}
+        return self.render_to_response(context)
+
+    def options(self, request, *args, **kwargs):
+        """
+        List all providers of current resource with their absolute urls.
+
+        Mapping of provider to url.
+        Provider -> URL
+
+        :param request: Request.
+        :return: Rendered response.
+        """
+        context = {name: self._get_provider_url(request, self.resource, name)
+                   for name, _, _, _ in self.providers}
         return self.render_to_response(context)
 
 
@@ -86,6 +161,8 @@ class ProviderAPIView(APIView):
             provider_module, provider_func = provider.rsplit('.', 1)
             module = importlib.import_module(provider_module)
             self.provider = getattr(module, provider_func, None)
+            if self.provider is None:
+                raise ValueError('Provider not found: %s' % (provider,))
         else:
             self.provider = provider